refactor: 핸들러 메서드 통합
