--- src/main/java/wooteco/subway/exception/IllegalLineArgumentException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.exception;
-
-public class IllegalLineArgumentException extends SubwayException {
-
-    public IllegalLineArgumentException(String message) {
-        super(message);
-    }
-}
