--- src/main/java/wooteco/subway/exception/SectionDeleteException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.exception;
-
-public class SectionDeleteException extends SubwayException {
-
-    public SectionDeleteException(String message) {
-        super(message);
-    }
-}
