--- src/main/java/wooteco/subway/exception/SectionUpdateException.java
@@ -1,7 +1,0 @@
-package wooteco.subway.exception;
-
-public class SectionUpdateException extends SubwayException {
-    public SectionUpdateException(String message) {
-        super(message);
-    }
-}
