+++ src/main/java/wooteco/subway/exception/illegalexception/IllegalException.java
@@ -1,0 +1,9 @@
+package wooteco.subway.exception.illegalexception;
+
+import wooteco.subway.exception.SubwayException;
+
+public class IllegalException extends SubwayException  {
+    public IllegalException(String message) {
+        super(message);
+    }
+}
