+++ src/main/java/wooteco/subway/exception/illegalexception/IllegalLineArgumentException.java
@@ -1,0 +1,8 @@
+package wooteco.subway.exception.illegalexception;
+
+public class IllegalLineArgumentException extends IllegalException {
+
+    public IllegalLineArgumentException(String message) {
+        super(message);
+    }
+}
