+++ src/main/java/wooteco/subway/exception/illegalexception/IllegalSectionArgumentException.java
@@ -1,6 +1,6 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.illegalexception;
 
-public class IllegalSectionArgumentException extends SubwayException {
+public class IllegalSectionArgumentException extends IllegalException {
 
     private static final String CREATE_SECTION_ERROR_MESSAGE = "잘못된 구간 정보입니다.";
