+++ src/main/java/wooteco/subway/exception/illegalexception/IllegalSectionStatusException.java
@@ -1,6 +1,6 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.illegalexception;
 
-public class IllegalSectionStatusException extends SubwayException {
+public class IllegalSectionStatusException extends IllegalException {
 
     private static final String SECTION_STATUS_ERROR_MESSAGE = "잘못된 구간 정보입니다.";
