+++ src/main/java/wooteco/subway/exception/sectionsexception/SectionDeleteException.java
@@ -1,0 +1,8 @@
+package wooteco.subway.exception.sectionsexception;
+
+public class SectionDeleteException extends SectionsException {
+
+    public SectionDeleteException(String message) {
+        super(message);
+    }
+}
