+++ src/main/java/wooteco/subway/exception/sectionsexception/SectionUpdateException.java
@@ -1,0 +1,7 @@
+package wooteco.subway.exception.sectionsexception;
+
+public class SectionUpdateException extends SectionsException {
+    public SectionUpdateException(String message) {
+        super(message);
+    }
+}
