+++ src/main/java/wooteco/subway/exception/sectionsexception/SectionsException.java
@@ -1,0 +1,10 @@
+package wooteco.subway.exception.sectionsexception;
+
+import wooteco.subway.exception.SubwayException;
+
+public class SectionsException extends SubwayException {
+
+    public SectionsException(String message) {
+        super(message);
+    }
+}
